feat: split hooks
- prep for hook refactor
