+++ lib/hooks/use-should-show-prompt.ts
@@ -1,33 +1,10 @@
-import { useLocalStorage } from 'usehooks-ts';
+import { usePromptStorage } from './use-prompt-storage.ts';
 
 type UseShouldShowPromptProps = {
   promptLocalStorageKey: string;
   promptOnVisit?: number;
   timesToShow?: number;
   withoutDefaults?: boolean;
-};
-
-export type PwaPromptData =
-  | {
-      isiOS: boolean; // represents whether or not the device is considered iphone/ipad
-      visits: number; // represents current prompt render (visit) count
-      dismissedAt?: number; // represents when the user dismissed the prompt as a UTC epoch
-    }
-  | undefined;
-
-export const usePromptStorage = (
-  promptLocalStorageKey: string,
-  withoutDefaults: boolean
-) => {
-  return useLocalStorage<PwaPromptData>(
-    promptLocalStorageKey,
-    !withoutDefaults
-      ? {
-          isiOS: false,
-          visits: 0,
-        }
-      : undefined
-  );
 };
 
 export const useShouldShowPrompt = ({
